add Gitpod config
Taken from eclipse-theia/theia and slightly adapted for this repo.
I have added the generation of preview and production packages
for Linux. This provides an easy means of testing and looking
under the hood, e.g. for troubleshooting, in an environment that
should offer reproducibility.

For enhanced benefit we can consider having webmaster enable the
Gitpod integration for this repo. Then we'd have automatic workspaces
built for PRs.

To open this PR in Gitpod and look-around, either use the browser
add-on or use this PR URL with the following string:
"https://gitpod.io/#" .

Once merged, this trick will work with any future branch/PR/commit
in this repo.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>